distribuindo as funcoes do crud
