+++ Intermediaria/1_CRUD/d_cr.h
@@ -1,0 +1,9 @@
+#include "struct_aluno.h"
+
+void data_create()
+{
+    char string[20];
+    printf("create-de-dentro-da-funcao\n");
+    scanf("%s", string);
+    printf("!");
+}
