+++ Intermediaria/1_CRUD/d_de.h
@@ -1,0 +1,6 @@
+#include "struct_aluno.h"
+
+void data_delete()
+{
+    printf("delete-de-dentro-da-funcao");
+}
