+++ Intermediaria/1_CRUD/d_re.h
@@ -1,0 +1,10 @@
+#include "struct_aluno.h"
+Aluno a;
+
+void data_read()
+{
+    
+    char string[20];
+    printf("pesquisa-de-dentro-da-funcao\n");
+    printf("%s", string);
+}
