+++ Intermediaria/1_CRUD/d_up.h
@@ -1,0 +1,6 @@
+#include "struct_aluno.h"
+
+void data_update()
+{
+    printf("update-de-dentro-da-funcao");
+}
